requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file dev.txt dev.in
#
alabaster==0.7.10         # via sphinx
appdirs==1.4.3            # via setuptools
args==0.1.0               # via clint
astroid==1.4.9            # via pylint
babel==2.3.4              # via sphinx
click==6.7                # via pip-tools
clint==0.5.1              # via twine
coverage==4.3.4
coveralls==1.1
django-classy-tags==0.8.0
django-cms==3.4.2
django-filer==1.2.7
django-formtools==2.0
django-mptt==0.8.7        # via django-filer
django-polymorphic==1.0.2  # via django-filer
django-sekizai==0.10.0
django-tools==0.32.9
django-treebeard==4.1.0
django==1.8.17
djangocms-admin-style==1.2.6.2
djangocms-text-ckeditor==3.4.0
docopt==0.6.2             # via coveralls
docutils==0.13.1          # via sphinx
easy-thumbnails==2.3      # via django-filer
first==2.0.1              # via pip-tools
flake8==3.3.0
html5lib==0.9999999
imagesize==0.7.1          # via sphinx
isort==4.2.5              # via pylint
jinja2==2.9.5             # via sphinx
lazy-object-proxy==1.2.2  # via astroid
markupsafe==1.0           # via jinja2
mccabe==0.6.1             # via flake8, pylint
olefile==0.44
packaging==16.8           # via setuptools
pillow==4.0.0
pip-tools==1.8.1
piprot==0.9.7
pkginfo==1.4.1            # via twine
pluggy==0.4.0             # via tox
py==1.4.33                # via pytest, tox
pycodestyle==2.3.1        # via flake8
pydocstyle==1.1.1
pyflakes==1.5.0           # via flake8
pygments==2.2.0           # via sphinx
pylint==1.6.5
pyparsing==2.2.0          # via packaging
pytest-cov==2.4.0
pytest-django==3.1.2
pytest==3.0.7
python-creole==1.3.1
pytz==2016.10             # via babel
requests-futures==0.9.7   # via piprot
requests-toolbelt==0.7.1  # via twine
requests==2.13.0          # via coveralls, piprot, requests-futures, requests-toolbelt, sphinx, twine
six==1.10.0
snowballstemmer==1.2.1    # via sphinx
sphinx==1.5.3
tox==2.6.0
twine==1.8.1
unidecode==0.4.20         # via django-filer
virtualenv==15.1.0        # via tox
wheel==0.29.0
wrapt==1.10.10            # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via pytest, twine
